Update functions.test.js
Added for loop tests for each of the basic four functions. This would equate to 1,600 tests run, but since they're single operation, they should take no time to process.

Also, the subtract function is misspelled in the original function file. Wonder what this would report if it can't find the function.